feat: page.tsx のロジックをカスタムフックに分離する
useSQLGeneration: SQL生成・データ加工ロジックを分離
useClipboardAndDownload: コピー・ダウンロードロジックを分離
page.tsx は両フックを呼び出す UI のみに簡略化。
docs/architecture.md のディレクトリ構成に hooks/ を追記。
